+++ spec/generators/rspec/mailer/mailer_generator_spec.rb
@@ -1,13 +1,9 @@
-require 'spec_helper'
-
 # Generators are not automatically loaded by Rails
 require 'generators/rspec/mailer/mailer_generator'
+require 'support/generators'
 
-describe Rspec::Generators::MailerGenerator, :type => :generator do
-  # Tell the generator where to put its output (what it thinks of as Rails.root)
-  destination File.expand_path("../../../../../tmp", __FILE__)
-
-  before { prepare_destination }
+RSpec.describe Rspec::Generators::MailerGenerator, :type => :generator do
+  setup_default_destination
 
   describe 'mailer spec' do
     subject { file('spec/mailers/posts_spec.rb') }